+++ src/template/string.rs
@@ -9,12 +9,15 @@ use alloc::boxed::Box;
 #[cfg(feature = "alloc")]
 use alloc::rc::Rc;
 #[cfg(feature = "alloc")]
+use alloc::string::String;
+#[cfg(feature = "alloc")]
 use alloc::sync::Arc;
 
 use crate::spec::Spec;
-use crate::template::context::Context;
+use crate::template::components::{VarListIter, VarName};
+use crate::template::context::{Context, DynamicContext};
 use crate::template::error::{Error, ErrorKind};
-use crate::template::expand::Expanded;
+use crate::template::expand::{expand_whole_dynamic, Chunk, Chunks, Expanded};
 use crate::template::parser::validate_template_str;
 
 #[cfg(feature = "alloc")]
@@ -79,7 +82,6 @@ mod owned;
 /// ```
 #[cfg_attr(feature = "serde", derive(serde::Serialize))]
 #[cfg_attr(feature = "serde", serde(transparent))]
-#[cfg_attr(docsrs, doc(cfg(feature = "alloc")))]
 #[derive(PartialEq, Eq, PartialOrd, Ord, Hash)]
 pub struct UriTemplateStr {
     /// The raw string.
@@ -257,6 +259,127 @@ impl UriTemplateStr {
     ) -> Result<Expanded<'a, S, C>, Error> {
         Expanded::new(self, context)
     }
+
+    /// Expands the template with the given dynamic context.
+    ///
+    #[cfg_attr(
+        feature = "alloc",
+        doc = concat!(
+            "If you need the allocated [`String`], use",
+            "[`expand_dynamic_to_string`][`Self::expand_dynamic_to_string`]."
+        )
+    )]
+    ///
+    /// See the documentation for [`DynamicContext`] for usage.
+    pub fn expand_dynamic<S: Spec, W: fmt::Write, C: DynamicContext>(
+        &self,
+        writer: &mut W,
+        context: &mut C,
+    ) -> Result<(), Error> {
+        expand_whole_dynamic::<S, _, _>(self, writer, context)
+    }
+
+    /// Expands the template into a string, with the given dynamic context.
+    ///
+    /// This is basically [`expand_dynamic`][`Self::expand_dynamic`] method
+    /// that returns an owned string instead of writing to the given writer.
+    ///
+    /// See the documentation for [`DynamicContext`] for usage.
+    ///
+    /// # Examples
+    ///
+    /// ```
+    /// # #[cfg(feature = "alloc")]
+    /// # extern crate alloc;
+    /// # use iri_string::template::Error;
+    /// # #[cfg(feature = "alloc")] {
+    /// # use alloc::string::String;
+    /// use iri_string::template::UriTemplateStr;
+    /// # use iri_string::template::context::{DynamicContext, Visitor, VisitPurpose};
+    /// use iri_string::spec::UriSpec;
+    ///
+    /// struct MyContext<'a> {
+    ///     // See the documentation for `DynamicContext`.
+    /// #     /// Target path.
+    /// #     target: &'a str,
+    /// #     /// Username.
+    /// #     username: Option<&'a str>,
+    /// #     /// A flag to remember whether the URI template
+    /// #     /// attempted to use `username` variable.
+    /// #     username_visited: bool,
+    /// }
+    /// #
+    /// # impl DynamicContext for MyContext<'_> {
+    /// #     fn on_expansion_start(&mut self) {
+    /// #         // Reset the state.
+    /// #         self.username_visited = false;
+    /// #     }
+    /// #     fn visit_dynamic<V: Visitor>(&mut self, visitor: V) -> V::Result {
+    /// #         match visitor.var_name().as_str() {
+    /// #             "target" => visitor.visit_string(self.target),
+    /// #             "username" => {
+    /// #                 if visitor.purpose() == VisitPurpose::Expand {
+    /// #                     // The variable `username` is being used
+    /// #                     // on the template expansion.
+    /// #                     // Don't care whether `username` is defined or not.
+    /// #                     self.username_visited = true;
+    /// #                 }
+    /// #                 if let Some(username) = &self.username {
+    /// #                     visitor.visit_string(username)
+    /// #                 } else {
+    /// #                     visitor.visit_undefined()
+    /// #                 }
+    /// #             }
+    /// #             _ => visitor.visit_undefined(),
+    /// #         }
+    /// #     }
+    /// # }
+    ///
+    /// let mut context = MyContext {
+    ///     target: "/posts/1",
+    ///     username: Some("the_admin"),
+    ///     username_visited: false,
+    /// };
+    ///
+    /// // No access to the variable `username`.
+    /// let template = UriTemplateStr::new("{+target}{?username}")?;
+    /// let s = template.expand_dynamic_to_string::<UriSpec, _>(&mut context)?;
+    /// assert_eq!(s, "/posts/1?username=the_admin");
+    /// assert!(context.username_visited);
+    /// # }
+    /// # Ok::<_, Error>(())
+    /// ```
+    #[cfg(feature = "alloc")]
+    pub fn expand_dynamic_to_string<S: Spec, C: DynamicContext>(
+        &self,
+        context: &mut C,
+    ) -> Result<String, Error> {
+        let mut buf = String::new();
+        expand_whole_dynamic::<S, _, _>(self, &mut buf, context)?;
+        Ok(buf)
+    }
+
+    /// Returns an iterator of variables in the template.
+    ///
+    /// # Examples
+    ///
+    /// ```
+    /// # use iri_string::template::Error;
+    /// use iri_string::template::UriTemplateStr;
+    ///
+    /// let template = UriTemplateStr::new("foo{/bar*,baz:4}{?qux}{&bar*}")?;
+    /// let mut vars = template.variables();
+    /// assert_eq!(vars.next().map(|var| var.as_str()), Some("bar"));
+    /// assert_eq!(vars.next().map(|var| var.as_str()), Some("baz"));
+    /// assert_eq!(vars.next().map(|var| var.as_str()), Some("qux"));
+    /// assert_eq!(vars.next().map(|var| var.as_str()), Some("bar"));
+    /// # Ok::<_, Error>(())
+    /// ```
+    #[inline]
+    #[must_use]
+    pub fn variables(&self) -> UriTemplateVariables<'_> {
+        UriTemplateVariables::new(self)
+    }
 }
 
 impl fmt::Debug for UriTemplateStr {
@@ -413,7 +536,6 @@ mod __serde_slice {
 
     // About `'de` and `'a`, see
     // <https://serde.rs/lifetimes.html#the-deserializede-lifetime>.
-    #[cfg_attr(docsrs, doc(cfg(feature = "serde")))]
     impl<'a, 'de: 'a> Deserialize<'de> for &'a UriTemplateStr {
         #[inline]
         fn deserialize<D>(deserializer: D) -> Result<Self, D::Error>
@@ -421,6 +543,104 @@ mod __serde_slice {
             D: Deserializer<'de>,
         {
             deserializer.deserialize_string(CustomStrVisitor)
+        }
+    }
+}
+
+/// An iterator of variables in a URI template.
+#[derive(Debug, Clone)]
+pub struct UriTemplateVariables<'a> {
+    /// Chunks iterator.
+    chunks: Chunks<'a>,
+    /// Variables in the last chunk.
+    vars_in_chunk: Option<VarListIter<'a>>,
+}
+
+impl<'a> UriTemplateVariables<'a> {
+    /// Creates a variables iterator from the URI template.
+    #[inline]
+    #[must_use]
+    fn new(template: &'a UriTemplateStr) -> Self {
+        Self {
+            chunks: Chunks::new(template),
+            vars_in_chunk: None,
+        }
+    }
+}
+
+impl<'a> Iterator for UriTemplateVariables<'a> {
+    type Item = VarName<'a>;
+
+    fn next(&mut self) -> Option<Self::Item> {
+        loop {
+            if let Some(vars) = &mut self.vars_in_chunk {
+                match vars.next() {
+                    Some((_len, spec)) => return Some(spec.name()),
+                    None => self.vars_in_chunk = None,
+                }
+            }
+            let expr = self.chunks.find_map(|chunk| match chunk {
+                Chunk::Literal(_) => None,
+                Chunk::Expr(v) => Some(v),
+            });
+            self.vars_in_chunk = match expr {
+                Some(expr) => Some(expr.decompose().1.into_iter()),
+                None => return None,
+            }
+        }
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+
+    use crate::spec::IriSpec;
+    use crate::template::context::{AssocVisitor, ListVisitor, Visitor};
+
+    struct TestContext;
+    impl Context for TestContext {
+        fn visit<V: Visitor>(&self, visitor: V) -> V::Result {
+            match visitor.var_name().as_str() {
+                "str" => visitor.visit_string("string"),
+                "list" => visitor
+                    .visit_list()
+                    .visit_items_and_finish(["item0", "item1", "item2"]),
+                "assoc" => visitor
+                    .visit_assoc()
+                    .visit_entries_and_finish([("key0", "value0"), ("key1", "value1")]),
+                _ => visitor.visit_undefined(),
+            }
+        }
+    }
+
+    #[test]
+    fn expand_error_pos() {
+        {
+            let e = UriTemplateStr::new("foo{list:4}")
+                .unwrap()
+                .expand::<IriSpec, _>(&TestContext)
+                .err()
+                .map(|e| e.location());
+            assert_eq!(e, Some("foo{".len()));
+        }
+
+        {
+            let e = UriTemplateStr::new("foo{/list*,list:4}")
+                .unwrap()
+                .expand::<IriSpec, _>(&TestContext)
+                .err()
+                .map(|e| e.location());
+            assert_eq!(e, Some("foo{/list*,".len()));
+        }
+
+        {
+            let e = UriTemplateStr::new("foo{/str:3,list*,assoc:4}")
+                .unwrap()
+                .expand::<IriSpec, _>(&TestContext)
+                .err()
+                .map(|e| e.location());
+            assert_eq!(e, Some("foo{/str:3,list*,".len()));
         }
     }
 }