+++ src/types/generic/absolute.rs
@@ -319,6 +319,18 @@ impl<S: Spec> RiAbsoluteStr<S> {
 
     /// Returns the normalized IRI.
     ///
+    /// # Notes
+    ///
+    /// For some abnormal IRIs, the normalization can produce semantically
+    /// incorrect string that looks syntactically valid. To avoid security
+    /// issues by this trap, the normalization algorithm by this crate
+    /// automatically applies the workaround.
+    ///
+    /// If you worry about this, test by
+    /// [`RiAbsoluteStr::ensure_rfc3986_normalizable`] method or
+    /// [`Normalized::ensure_rfc3986_normalizable`] before using the result
+    /// string.
+    ///
     /// # Examples
     ///
     /// ```